requirements.txt
numpy~=1.26.4
matplotlib~=3.8.4
scikit-image~=0.23.2
filterpy~=1.4.5
opencv-python~=4.9.0.80
imutils~=0.5.4
requests~=2.25.1
ultralytics~=8.2.2
future~=0.18.2
pillow~=10.3.0
pytesseract~=0.3.10
easyocr~=1.7.1